circle component added
